# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI-Powered Internship Recommendation Engine API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /recommendations/{studentId}:
    get:
      operationId: getRecommendations
      tags: [recommendations]
      summary: Get AI-powered ranked internship recommendations for a student
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RecommendationResult"
  /recommendations/{studentId}/career-paths:
    get:
      operationId: getCareerPaths
      tags: [recommendations]
      summary: Get AI-generated career path recommendations for a student
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CareerPath"
...
components:
  schemas:
...
    Internship:
      type: object
      required: [id, title, company, domain, location, requiredSkills, description, duration, stipend, applicationDeadline, popularity, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        company:
          type: string
        domain:
          type: string
        location:
          type: string
        requiredSkills:
          type: array
          items:
            type: string
        description:
          type: string
        duration:
          type: string
        stipend:
          type: string
        applicationDeadline:
          type: string
        popularity:
          type: integer
        createdAt:
          type: string
...
    RecommendationResult:
      type: object
      required: [internship, score, skillMatchScore, interestScore, locationScore, popularityScore, matchedSkills, missingSkills, aiInsight]
      properties:
        internship:
          $ref: "#/components/schemas/Internship"
        score:
          type: number
        skillMatchScore:
          type: number
        interestScore:
          type: number
        locationScore:
          type: number
        popularityScore:
          type: number
        matchedSkills:
          type: array
          items:
            type: string
        missingSkills:
          type: array
          items:
            type: string
        aiInsight:
          type: string
    CareerPath:
      type: object
      required: [title, description, requiredSkills, timelineMonths, matchScore, steps]
      properties:
        title:
          type: string
        description:
          type: string
        requiredSkills:
          type: array
          items:
            type: string
        timelineMonths:
          type: integer
        matchScore:
          type: number
        steps:
          type: array
          items:
            type: string
...